+++ examples/stencil-fetch/main.ts
@@ -1,4 +1,0 @@
-import "../vanilla/fetch/my-fetch.js";
-import { defineCustomElement } from "../stencil-components/dist/components/stencil-fetch-app.js";
-
-defineCustomElement();