+++ examples/stencil/counter/main.ts
@@ -1,0 +1,5 @@
+import "../../vanilla/counter/my-counter.js";
+import { defineCustomElement } from "../components/dist/components/stencil-counter-app.js";
+
+defineCustomElement();
+