+++ examples/stencil/fetch/main.ts
@@ -1,0 +1,5 @@
+import "../../vanilla/fetch/my-fetch.js";
+import { defineCustomElement } from "../components/dist/components/stencil-fetch-app.js";
+
+defineCustomElement();
+